Update Welcome page links behavior
This commit updates the welcome page behavior to change links
based on whether or not a user is signed in and adds a dropdown
menu to the top of the page for signed in users. In order for
the drop down to work, jQuery and bootstrap were required and
installed with yarn.