Implement tests to set item category for LineItem
Item category is an important field for the LineItem object.
It will help us later determine if it is a taxable item or not.
This commit implements the tests for this field that utilize two models.
The BasketReader model which will parse the uploaded text file and the
LineItems::ItemCategory model which will return the item category.
Tests include the expected behavior of each model as well as a test to
ensure that the item category is set correctly under LineItem.