Implement testing for creating a basket
These tests define the behavior of the baskets controller when
creating a basket. Upon successful creation, the user should be
redirected to the basket show page where the final totals will be
displayed along with the line items. It is important to check sad
paths as well when the user tries to upload an incorrect file type
or when the file cannot be parsed or bad item generation.