Add ReceiptManager module
This module is responsible for managing the connections between the
necessary models for ReceiptManager and other models that you need
to have a full application, such as User or Product. This allows for
developers to easily include the ReceiptManager logic into their code
and that if their 'User' or 'Product' models have different names
that they can easily override the constants that store these.

Without this, it would mean they would need to overwrite multiple
methods that reference these models. This would cause more work as
well as more difficult to update ReceiptManager logic when there are
updates.