Add Receipt logic with tests
Note, this model contains the method self#create_from_file which acts
as a new way to create Receipt instances. Because of this, this commit
also adds an instance method to User to allow creating of receipts by
a receipt file in the Rails idiomatic way. This means that calling the
create_receipt_from_file method on a user instance will pass the user
automatically over to Receipt.new without having to give it as an
argument. This is to encourage the code to be more the Rails idiomatic.