Refactor create action for baskets controller
Refactored to improve the readability of the code. Basket_items_file
renamed to just basket_file to make it more clear. Extracting the
BasketHelper to add_items method works better as a model method and
removes that logic from the controller. Relocate file_type_allowed to
the basket helper as it should know whether the file is allowed or not.
Changes file_type_allowed to file_type_not_allowed to work with and if
instead of else to make it more clear.