Add ReceiptsController#show
This page shows the user an individual receipt. You must be logged
in otherwise you will be redirected to the homepage. You also cannot
see receipts that do not belong to you. This is done implicity through
searching for the receipt (by the given id) on the User's receipts.
If nothing is found, you will be redirected to the homepage. Integrated
with tests.

A new method was also added to the ApplicationHelper which
is #money_as_string. Pass it over any number and it will make it a
two decimal place string. Useful for wanting to display money in views.